Remove Masters, Remove Comments and Added Auto Semester
As messages suggests. Current logontosussed will not work due to python if statements being awkward. quickstart hasn't changed much and now might be able to send JSON files to the web server. I don't think it can take them yet so its worth commenting that out it is certain that it can work.